move file upload to kitchen sink demo
