+++ examples/kitchen-sink/composer.json
@@ -1,0 +1,18 @@
+{
+  "name": "bunnycdn/storage-example-file-upload",
+  "description": "File upload form example for Bunny Storage SDK",
+  "type": "project",
+  "require": {
+    "php": ">=8.1",
+    "bunnycdn/storage": "*"
+  },
+  "repositories": [
+    {
+      "type": "path",
+      "url": "../../"
+    }
+  ],
+  "scripts": {
+    "start": "php -S localhost:8000 index.php"
+  }
+}
